clean up and reorganize files
