+++ restapi/api/device_endpoints.go
@@ -7,21 +7,7 @@ import (
 	"github.com/gin-gonic/gin"
 	log "github.com/sirupsen/logrus"
 	"net/http"
-	"strings"
 )
-
-func registerDeviceSpecificEndpoints(router *gin.RouterGroup) {
-	device := router.Group("/device/:udid")
-	device.Use(DeviceMiddleware())
-	device.GET("/info", Info)
-	device.GET("/screenshot", Screenshot)
-
-	initAppRoutes(device)
-
-	streamingDevice := device.Group("")
-	streamingDevice.Use(HeadersMiddleware())
-	streamingDevice.GET("/syslog", Syslog)
-}
 
 // Info gets device info
 // Info                godoc
@@ -78,31 +64,3 @@ func Screenshot(c *gin.Context) {
 	c.Header("Content-Type", "image/png")
 	c.Data(http.StatusOK, "application/octet-stream", b)
 }
-
-// DeviceMiddleware makes sure a udid was specified and that a device with that UDID
-// is connected with the host. Will return 404 if the device is not found or 500 if something
-// else went wrong. Use `device := c.MustGet(IOS_KEY).(ios.DeviceEntry)` to acquire the device
-// in downstream handlers.
-func DeviceMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		udid := c.Param("udid")
-
-		if udid == "" {
-			c.AbortWithStatusJSON(http.StatusUnprocessableEntity, gin.H{"message": "udid is missing"})
-			return
-		}
-		device, err := ios.GetDevice(udid)
-		if err != nil {
-			if strings.Contains(err.Error(), "not found") {
-				c.AbortWithStatusJSON(http.StatusNotFound, gin.H{"message": "device not found on the host"})
-				return
-			}
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err})
-			return
-		}
-		c.Set(IOS_KEY, device)
-		c.Next()
-	}
-}
-
-const IOS_KEY = "go_ios_device"
